backend/requirements.txt
# requirements.txt - COMPLETE LIST

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI Engine Dependencies
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.13
langgraph==0.0.20

# Google Gemini (NEW SDK)
google-genai==0.3.0

# Vector Database
chromadb==0.4.22

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Configuration
python-dotenv==1.0.0

# Database (for sessions - optional initially)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring
python-json-logger==2.0.7

# Additional utilities
requests==2.31.0
aiohttp==3.9.1